Fixed frequent crashes in The Witness. Apparently two entries in GOAL-TABLES were disabled because of compilation errors, and then never re-enabled. I'm guessing this caused it to read outside of the table, causing it to use invalid data.
I-LINDER and I-CAT were not defined anywhere in the code. They're in the
GOAL-FUNCTION field of the table. I believe this specifies a routine to
call whenever an actor reaches a set goal. Apparently Mr. Linder doesn't
need this mechanism to move to his office, and the cat never does much
of anything.

In the original witnessdat.zap, they are still defined as 0 under the
"USER DEFINED CONSTANTS ARE INITIALIZED HERE" heading. Did the original
compiler do that automatically for missing identifiers?

It seems to me that the least intrusive fix is to re-enable the two
missing entries, and define I-LINDER and I-CAT as constants with the
value 0. That should be equivalent to the old behavior, and the table
remains easily understandable.

Doing this, I've been able to play the game to its conclusion (though I
skipped all of the optional clues, and I did run into a parser issue
near the end), so it's a big improvement.