Reverted OUTLETS-F to an earlier version. Apparently, at some point between releases 77 and 79, Infocom's compiler changed from automatically removing one space after newline ('|') to not doing it. That would put the change some time between August 14 and November 22, 1988.
As a consequence, in Release 77 each line in the list of communications
outlets is prefixed by three spaces. In Release 79 the first line is
prefixed with three spaces, and the remaining ones with four. At some
point after that, the function was changed to add a fourth space to the
first row.

ZILF's default behavior is to remove the extra space, so it will
prefix the first row with four spaces, and the remaining ones with
three.

Since using <SETG PRESERVE-SPACES? T> will break some other text
formatting in Library Mode, I think the best change is to just remove
the extra space in OUTLETS-F. That mimics the behavior in Release 77 and
all surviving test versions before that.