Adjusted the points to get the total to 400.
The difference in scoring between the two games is that the full game
awards 5 points for solving the diamond maze and 3 points for reaching
Crypt Anteroom.

Mini-Zork II doesn't have either of these. It compensated by awarding 10
points for shrinking in the Alice area, but did not award 10 points for
reaching the Cerberus Room.

This made the maximum points in Mini-Zork II 400 + 10 - 8 = 392 points.

I've compensated by adding 10 points for reaching the Cerberus Room, and
lowering the points for shrinking in the Alice area (where the full game
awards no points) from 10 to 8 points.

That should give the correct maximum score. I'll verify it next time I
play through the whole thing.