Fixed balloon location after takeoff.
I don't fully understand this, but this is what the original game did
unless the room had been destroyed by an explosion. Mini-Zork II handles
the explosion differently, so MUNGBIT is never set. I guess the
assignment was removed accidentally when that was rewritten.

Anyway, this seems to work fine now.