Some more score balancing
I've removed the points for reaching Cerberus again, since it's no
longer a puzzle. (-10)

I've reduced the points for shrinking down in the Alice area from 8 (was
10) to 5, since the game mostly uses multiples of 5 for scoring. (-3)

I've added 10 points for taking off in the balloon. (+10)

Reaching the Crypt is now worth 5 points (was 2) (+3).

That still balances out to 400 points, and makes sense to me.