Fixed the description of the menhir when floating in the air.
Before it would use MENHIR-POSITION 3 both for when the demon had moved
the menhir to the pentagram room and when it was floating in the air.
Now it uses 2 for all cases where the menhir is no longer in the Menhir
Room, since both these cases have should have the same description. 3 is
now only used for when it's floating.