Fixed menhir being visible even after being removed.
According to the ZIL manual, local-globals have to be inside
LOCAL-GLOBALS, and be listed in the room's GLOBAL property to be
visible. However, it looks to me as if any object in GLOBAL would be
found, regardless of where it was physically located.

This breaks the Zork II menhir, which is a local-global object (visible
from both the Menhir Room and the Kennel), but can still be both removed
from the game (if you tell the demon to take it) and moved to the
Pentagram Room (if you ask the demon to give it to you).

This bug happens in the full game as well, and if this really is the
correct bugfix it's probably in a bunch of other games as well. Though
it's quite possible that Zork II is the only game that ever (re)moves a
local-global object like this.