Test the location of a burning object before removing it, not after.
This restores the obviously intended behavior where burning an object
while carrying it (or sitting inside it, but I don't think that ever
happens in Zork II?) is fatal. It's cheap, it's cruel, it's Zork.

Incidentally, burning the fuse while carrying it is not fatal, since
that's handled by the fuse's action routine, not the default one.