Fixed menhir description when it's floating in the air.
The game used MENHIR-POSITION 2 to indicate that the demon had removed
the menhir from the game, and 3 to indicate that the demon had moved the
menhir to the Pentagram Room. It also used position 3 to indicate that
the menhir was being levitated by the wand.

That meant that the description of the levitating menhir was never
printed.

Since the menhir should be described the same way in the Menhir Room,
whether the demon removed the menhir or moved it to another room, it now
uses position 2 for both these cases, and position 3 prints the
desciption of the levitating menhir.