Don't automatically take the fuse on "LIGHT FUSE"
This bug happened because the "LIGHT OBJECT" syntax specified both TAKE
and HAVE for the object to light. Since the verb function is V-LAMP-ON,
I assume it's primarily intended for the lamp, but at some point someone
decided - correctly, I think - that "LIGHT FUSE" is a reasonably
command.

What's not so reasonable is that the parser then automatically took the
fuse out of the brick, leaving your looking like Wile E. Coyote, holding
just the burning string and with not enough time to put it back into the
brick again and make it out alive.

By giving the fuse the TAKEBIT flag, the parser no longer tries to
automatically take it. And by leaving out the HAVE flag on the syntax,
the parser no longer requires you to be holding the object.

I can't think of anything that would break from this change. Worst case
would be that you manage to light the lamp without carrying it. But if
that's a problem, then "TURN ON LAMP" already triggers the same problem
since that syntax has neither HAVE nor TAKE. (Maybe to make sure that
"TURN ON BIG IMMOVABLE MACHINE" still works?)