Fixed fighting with the thief.
This bug was caused by faulty end-of-table checks for the VILLAINS
table. An LTABLE is a 1-based list, with the number of elements stored
at index 0. Therefore, only exit the loop after going *past* LEN.

This bug seems to be present in the full Zork I as well, but in that
game there is a third entry in the table for the cyclops. Which is
interesting, because you were able to fight the cyclops in the MDL
version of Zork, and Zork I does contain the messages for fighting the
cyclops. Hmmm...