Fixed double room description after being dropped by the giant bat.
This fix is heavily based on the full Zork I. I was tempted to trigger
the bat on M-END instead of M-ENTER, since then the room descriptions
would have been printed correctly without any "cheating".

However, in this version (and only this version?), one of the rooms
where the bay may drop you is the Gas Room. And then the explosion if
you are carrying the torch isn't triggered until the next move.

This way, the Gas Room works as intended.