Make room turn dark if the thief leaves with the only light.
The thief may still be in the room, albeit invisible, so his inventory
has to be made invisible too for the light to go away. This will happen
if you give him your light source, and I think this bug happens in the
full game as well.

The thief stealing the only light is a different matter. That happens as
he's leaving, and anything he steals automatically becomes invisible. I
don't think that ever happens in this version, since the torch has
SACREDBIT.